+ recently used currencies are at the top
Show "last" currencies at the top of the currency list

Problem was
----------
For many users, this behavior is challenging, since they only use a couple currencies within the app (eg only USD and EUR). By showing the full list of ~100 currencies, it forces them to scroll far through the list to view this every time.